+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Bugfix: Update set of items that are never kept on dangerous deaths. (#364)
 - Bugfix: Don't report inaccurate completed collections count, when character summary tab was not selected. (#374)
 - Bugfix: Avoid undercounting diary completions for notifications that occur shortly after a teleport. (#373)
+- Dev: Small refactor of leagues notifier. (#387)
 - Dev: Execute screenshot rescaling on background thread to minimize FPS impact. (#378)
 
 ## 1.7.2